`out_of_stock` showing directly from server
on the frontend in index.html we have  a condition available_stock not greater than 0 then show out of stock.
But due to caching on the mobile app it is not worked. Hence directly from server showing out of stock for available_stock which are 0 by manually typed column names